Refactoring
- resolves secvisogram/secvisogram#202
- rename function to match test number
- move file to match structure
